FN-5668

+++ focDataSource_DB/src/com/foc/focDataSourceDB/db/DBManagerServer.java
@@ -296,7 +296,7 @@ public class DBManagerServer {
 			    	  	}
 			    	  	
 			    	  	while(rs.next()){
-			    	  		int ref = rs.getInt(indexOfREF);
+			    	  		long ref = rs.getLong(indexOfREF);
 	
 			    	  		focObject.setReference_WithFocListRefHashMapAdjustment(ref);
 			    	  	}
